:rotating_light: Rename views/Home.vue to pages/HomePage.vue
